pyproject.toml
[project]
name = "hubspot-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.10.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
